+++ src/macos/PeripheralBase.mm
@@ -16,6 +16,7 @@ PeripheralBase::PeripheralBase(void* opaque_peripheral, void* opaque_adapter, ad
     is_connectable_ = advertising_data.connectable;
     manual_disconnect_triggered_ = false;
     manufacturer_data_ = advertising_data.manufacturer_data;
+    rssi_ = advertising_data.rssi;
 }
 
 PeripheralBase::~PeripheralBase() {
@@ -34,6 +35,16 @@ std::string PeripheralBase::identifier() {
 BluetoothAddress PeripheralBase::address() {
     PeripheralBaseMacOS* internal = (__bridge PeripheralBaseMacOS*)opaque_internal_;
     return std::string([[internal address] UTF8String]);
+}
+
+int16_t PeripheralBase::rssi() {
+    return rssi_;
+}
+
+void PeripheralBase::update_advertising_data(advertising_data_t advertising_data) {
+    is_connectable_ = advertising_data.connectable;
+    manufacturer_data_ = advertising_data.manufacturer_data;
+    rssi_ = advertising_data.rssi;
 }
 
 void PeripheralBase::connect() {