+++ src_c/peripheral.cpp
@@ -6,6 +6,7 @@
 
 #include <cstring>
 #include <map>
+#include <climits>
 
 void simpleble_peripheral_release_handle(simpleble_peripheral_t handle) {
     if (handle == nullptr) {
@@ -38,6 +39,15 @@ char* simpleble_peripheral_address(simpleble_peripheral_t handle) {
     char* c_address = (char*)malloc(address.size() + 1);
     strcpy(c_address, address.c_str());
     return c_address;
+}
+
+int16_t simpleble_peripheral_rssi(simpleble_peripheral_t handle) {
+    if (handle == nullptr) {
+        return INT16_MIN;
+    }
+
+    SimpleBLE::Safe::Peripheral* peripheral = (SimpleBLE::Safe::Peripheral*)handle;
+    return peripheral->rssi().value_or(INT16_MIN);
 }
 
 simpleble_err_t simpleble_peripheral_connect(simpleble_peripheral_t handle) {