+++ docs/changelog.rst
@@ -6,7 +6,18 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog`_, and this project adheres to
 `Semantic Versioning`_.
 
-[0.3.0] - 2021-04-03
+[0.4.0] - 2022-XX-XX
+--------------------
+
+**Added**
+
+-  Expose RSSI as a property of ``SimpleBLE::Peripheral``.
+
+**Changed**
+
+-  Updated Linux implementation to use SimpleBluez v0.4.0.
+
+[0.3.0] - 2022-04-03
 --------------------
 
 **Added**
@@ -24,7 +35,7 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 -  Updated Linux implementation to use SimpleBluez v0.3.1.
 -  Migrated to using safe callbacks from external vendor (kvn::safe_callback).
 
-[0.2.0] - 2021-02-13
+[0.2.0] - 2022-02-13
 --------------------
 
 **Added**